Initial comment of already working code.
This project was developed to explore the ways in which regression based
estimations problem may be solved using machine learning. Initially,
simple and polynomial regression was computed for a compplete
sub-selection of possible parameters. In addition to this, the
NNTestingFramework presents a sophisticated grid-search like
functionality that was used to test a larger number of different
feed-forward neural networks.

The framework automatically plots graphs and writes logs, and ranks all
of the FFNN on unseen data in an appropriate manner, such as to minimise
the inherit bias.

Short scripts testing other methods are also present.